Add game functionality
This commit adds the ability to start a game and answer questions.

The game is started by clicking the "Play on all categories" button on the
home page. The user is then redirected to the game page where they can answer
questions.

The game page shows the user's progress through the game and the current
question. The user can answer the question by selecting one of the answers.

The game page also shows the user's score.