Add time limit to game questions
This commit adds a time limit to game questions. The time limit is set to 30
seconds by default, but can be changed by setting the `time_limit` attribute of
the `GameQuestion` object.

The time remaining is displayed to the user, and the user is prevented from
submitting an answer after the time limit has expired.